Update: now logger uses project utils
