+++ CLogger/CLogger.vcxproj
@@ -189,7 +189,6 @@
     <ClInclude Include="include\ELogLevel\ELogLevel.h" />
     <ClInclude Include="include\Log.h" />
     <ClInclude Include="include\stdafx.h" />
-    <ClInclude Include="include\Utils\Utils.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\PropertySheets\CodeStylerSettings\.clang-format" />
